Rename file for coherence across projects and main purpose #10310
